Download elliptic curve cryptography project
The project implements the El Gamal cipher based on elliptic curves. Three arguments are submitted to the input - the name of the encrypted file, the file that will contain the cipher and the one that will contain the decrypted data.  Accordingly, an encrypted file is obtained at the output, and its decryption is implemented.